Update PR based on code review feedback.
