+++ util-hadoop/src/main/hadoop3/java/com/google/cloud/hadoop/util/ConfigurationEntriesAdapterSpecific.java
@@ -14,16 +14,22 @@
 
 package com.google.cloud.hadoop.util;
 
+import java.io.IOException;
 import org.apache.hadoop.conf.Configuration;
 
-public class ConfigurationEntriesAdapter extends ConfigurationEntriesAdapterGeneric {
+abstract class ConfigurationEntriesAdapterSpecific extends ConfigurationEntriesAdapterGeneric {
 
-  public ConfigurationEntriesAdapter(Configuration config) {
+  public ConfigurationEntriesAdapterSpecific(Configuration config) {
     super(config);
   }
 
   @Override
   public String getPassword(String key) {
-    return get(key);
+    try {
+      char[] val = config.getPassword(key);
+      return val == null ? null : String.valueOf(val);
+    } catch (IOException e) {
+      throw new RuntimeException(e);
+    }
   }
 }
